tests: breakout selenium.go from tests.go
